Removed limit, added paged search and attributes
To drastically reduce the sent data the attributes which are returned can
be specified now.

Furthermore a paged search is used because this also greatly improves
performance.

Page size is arbitrarily set to 300 since this seemed to be a "sweet spot"
and resulted in best performance when accessing the eth ldap