Update version, clean up TLS.
For some reason, start_tls would fail for the ldaps servers
(which use an ssl socket by default).

The socket is opened here:
https://github.com/cannatag/ldap3/blob/dev/ldap3/strategy/base.py#L134

And server.ssl is set to true for any ldaps address:
https://github.com/cannatag/ldap3/blob/dev/ldap3/core/server.py#L94

No idea why this did not crash before.
But now it does :(

Also updated version (but the new version does not touch the TLS code..)